Filter Movie Method

+++ MoviesApi/Dto/MovieDetailsDto.cs
@@ -5,7 +5,7 @@
         public int ID { get;set;}
         public string Title { get;set;}
         public int Year { get;set;}
-        public byte Rate { get;set;}
+        public double Rate { get;set;}
         public string StoreLine { get;set;}
         public string PosterUrl { get;set;}
         public byte GenreId { get;set;}
